+++ Tests/XcodeProjTests/Scheme/XCSchemeTests.swift
@@ -245,6 +245,28 @@ final class XCSchemeIntegrationTests: XCTestCase {
         XCTAssertNotEqual(runnableA1, remoteRunnableA1)
     }
 
+    func test_buildAction_runPostActionsOnFailure() throws {
+        // Given / When
+        let subject = try XCScheme(path: runPostActionsOnFailureSchemePath)
+
+        // Then
+        let buildAction = try XCTUnwrap(subject.buildAction)
+        XCTAssertTrue(buildAction.runPostActionsOnFailure == true)
+    }
+
+    func test_buildAction_runPostActionsOnFailure_serializingAndDeserializing() throws {
+        // Given
+        let scheme = try XCScheme(path: runPostActionsOnFailureSchemePath)
+        let subject = try XCTUnwrap(scheme.buildAction)
+
+        // When
+        let xml = subject.xmlElement()
+        let reconstructedSubject = try XCScheme.BuildAction(element: xml)
+
+        // Then
+        XCTAssertEqual(reconstructedSubject, subject)
+    }
+
     // MARK: - Private
 
     private func assert(scheme: XCScheme) {
@@ -253,7 +275,7 @@ final class XCSchemeIntegrationTests: XCTestCase {
         // Build action
         XCTAssertTrue(scheme.buildAction?.parallelizeBuild == true)
         XCTAssertTrue(scheme.buildAction?.buildImplicitDependencies == true)
-        XCTAssertTrue(scheme.buildAction?.runPostActionsOnFailure == false)
+        XCTAssertNil(scheme.buildAction?.runPostActionsOnFailure)
         XCTAssertTrue(scheme.buildAction?.buildActionEntries.first?.buildFor.contains(.testing) == true)
         XCTAssertTrue(scheme.buildAction?.buildActionEntries.first?.buildFor.contains(.running) == true)
         XCTAssertTrue(scheme.buildAction?.buildActionEntries.first?.buildFor.contains(.profiling) == true)
@@ -420,7 +442,7 @@ final class XCSchemeIntegrationTests: XCTestCase {
         // Build action
         XCTAssertTrue(scheme.buildAction?.parallelizeBuild == true)
         XCTAssertTrue(scheme.buildAction?.buildImplicitDependencies == true)
-        XCTAssertTrue(scheme.buildAction?.runPostActionsOnFailure == false)
+        XCTAssertNil(scheme.buildAction?.runPostActionsOnFailure)
         XCTAssertTrue(scheme.buildAction?.buildActionEntries.first?.buildFor.contains(.testing) == true)
         XCTAssertTrue(scheme.buildAction?.buildActionEntries.first?.buildFor.contains(.running) == false)
         XCTAssertTrue(scheme.buildAction?.buildActionEntries.first?.buildFor.contains(.profiling) == true)
@@ -525,5 +547,10 @@ final class XCSchemeIntegrationTests: XCTestCase {
 
     private var watchAppSchemePath: Path {
         fixturesPath() + "iOS/AppWithExtensions/AppWithExtensions.xcodeproj/xcshareddata/xcschemes/WatchApp.xcscheme"
+    }
+
+    private var runPostActionsOnFailureSchemePath: Path {
+        // A scheme with the `runPostActionsOnFailure` enabled
+        fixturesPath() + "Schemes/RunPostActionsOnFailure.xcscheme"
     }
 }