+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+## 8.0.0 - Amor
 ### Fixed
 
 - Adding group set incorrect parent in case of complex path [#614](https://github.com/tuist/XcodeProj/pull/614) by [@avdyushin](https://github.com/avdyushin)